wip: initial refactoring
